backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
groq==0.11.0
langchain==0.3.13
langchain-groq==0.2.3
httpx==0.27.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.9
pydantic-settings==2.5.2